% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{add_id_pixel_and_groups}
\alias{add_id_pixel_and_groups}
\title{Add id_pixel and kmeans groups from a reference dataframe. Make a left join.}
\usage{
add_id_pixel_and_groups(dataframe_list, reference_dataframe,
  join_by = c("lon", "lat"))
}
\arguments{
\item{dataframe_list}{a list of dplyr dataframes}

\item{reference_dataframe}{reference dataframe containing lon, lat, id_pixel, and groups.}

\item{join_by}{variables to join by}
}
\description{
For each dataframe in the data_frame list, id_pixel and the group from the kmeans analysis
is added. Note that the reference dataframe must contain longitude, latitude (lon and lat),
id_pixel and groups.
}
