% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_calculation_functions.R
\name{aggregate_stats}
\alias{aggregate_stats}
\title{Calculate multiple statistics on aggregated data}
\usage{
aggregate_stats(data, variable = "CHL1_mean", groups = list("id.pixel",
  "id.date"), stat_funs = list(avg = "mean", sdev = "sd"))
}
\arguments{
\item{data}{a dplyr data frame}

\item{groups}{variables to group by. A list.}

\item{stat_funs}{A list of functions to calculate the statistics with.
and returns a single value. For instance: mean, var, sd.}

\item{variables}{variable to use in the calculation of the statistic. A list with names.}
}
\value{
A list of dplyr dataframes.
}
\description{
This function groups the observations and calculates a given number of statistics.
The function returns a list of dataframes (ie a dataframe per statistic)
}

