% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-functions.R
\name{apply_mask}
\alias{apply_mask}
\title{Apply the mask to the data and remove earth pixels}
\usage{
apply_mask(data, mask)
}
\arguments{
\item{data}{data to be used. A data frame object.}

\item{mask}{a mask dataframe loaded using the load_mask function or the resize_mask function.}
}
\value{
a dplyr dataframe
}
\description{
Apply the mask to the data and remove earth pixels
}

