% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{approximate_NAs}
\alias{approximate_NAs}
\title{Approximate NAs}
\usage{
approximate_NAs(data, exclude_variables = list("lon", "lat", "id_pixel"))
}
\arguments{
\item{data}{a dplyr dataframe containing the data to be standardized}

\item{exclude_variables}{variables to be excluded from the standardization process. A list of characters.}
}
\value{
A dplyr dataframe
}
\description{
This function is used to fill missing NA values using common imputation methods.
}

