% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-data-functions.R
\name{assign_id_and_melt}
\alias{assign_id_and_melt}
\title{Given a list of data loaded using the load_all_as_list() function, this function is used to
assign a unique id to each pixel and bind all the rows together in a single dataframe.}
\usage{
assign_id_and_melt(data_list, coordinates = c("lon", "lat"))
}
\arguments{
\item{data_list}{A list of dplyr dataframes returned by the load_all_as_list function.}

\item{coordinates}{Unique identifier to be used in the id assignment process.}
}
\value{
A dplyr dataframe
}
\description{
Given a list of data loaded using the load_all_as_list() function, this function is used to
assign a unique id to each pixel and bind all the rows together in a single dataframe.
}

