% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_calculation_functions.R
\name{calculate_aggregate_stat}
\alias{calculate_aggregate_stat}
\title{Extract a single variable from a .nc file}
\usage{
calculate_aggregate_stat(data, variable = "CHL1_mean",
  groups = list("id.pixel", "id.date"), stat_fun = "mean", name = "avg")
}
\arguments{
\item{data}{a dplyr data frame}

\item{variable}{variable to use in the calculation of the statistic}

\item{groups}{variables to group by. A list.}

\item{stat_fun}{Function to calculate the statistic. This parameter must be a function that accepts a vector as an argument
and returns a single value. For instance: mean, var, sd.}

\item{name}{name to assign to the statistic}
}
\value{
A dplyr data frame.
}
\description{
This function groups the observations and calculates a given statistic.
Note: By default the function calculates the mean.
}

