% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop_selected_area}
\alias{crop_selected_area}
\title{This function is used to crop a selected area from a list of dataframes of
geographical observations}
\usage{
crop_selected_area(df_list, lower_left_lat_lon, upper_right_lat_lon)
}
\arguments{
\item{df_list}{a list of dplyr dataframe}

\item{lower_left_lat_lon}{lower left corner latitude and longitude of the selected area. Numeric vector.
Example c(52.00, -65.00)}

\item{upper_right_lat_lon}{upper right corner latitude and longitude of the selected area.
Numeric vector. Example: c(67.00, -42.00)}
}
\value{
Returns a list of dplyr dataframes
}
\description{
This function is used to crop a selected area from a list of dataframes of
geographical observations
}
