% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{export_data}
\alias{export_data}
\title{Extract results from the kmeans analysis}
\usage{
export_data(x, nc_dataframe, final_kmeans_results, filename,
  export_vars = list("lon", "lat", "id_pixel", "gruppo"),
  join_by = c("id_pixel", "lon", "lat"), unique_id = list("lon", "lat",
  "id_pixel"))
}
\arguments{
\item{x}{dataframe used in the kmeans analysis (prior to standardization)}

\item{nc_dataframe}{main dataframe}

\item{final_kmeans_results}{list of the final results obtained from the function extract_results}

\item{filename}{name of the .csv file to export}

\item{export_vars}{variables to export. A list of characters}

\item{join_by}{variables to join by. A vector of characters}

\item{unique_id}{a list of unique identifiers}
}
\description{
This function is used to extract the best results from the kmeans analysis.
The best result is considered to be the result characterized by the highest
Calinski-Harabasz index.
}
