% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{extract_results}
\alias{extract_results}
\title{Extract results from the kmeans analysis}
\usage{
extract_results(kmeans_results)
}
\arguments{
\item{kmeans_results}{a list of results obtained from the function kmeans_analysis}
}
\value{
A list with the best results (as defined above) of the kmeans analysis.
}
\description{
This function is used to extract the best results from the kmeans analysis.
The best result is considered to be the result characterized by the highest
Calinski-Harabasz index.
}

