% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_single_date}
\alias{filter_single_date}
\title{Filter a given date within a given interval:
If the date is within the interval, return the date
If not, return NULL.}
\usage{
filter_single_date(current_date, from, to)
}
\arguments{
\item{current_date}{date to be filtered. A Date object.}

\item{from}{Initial date (included). A Date object.}

\item{to}{Final date (included). A Date object.}
}
\value{
Returns NULL if the date is out of the given range, otherwise returns the current_date parameter
}
\description{
Filter a given date within a given interval:
If the date is within the interval, return the date
If not, return NULL.
}
