% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{fit_random_forest}
\alias{fit_random_forest}
\title{Fit a random forest model}
\usage{
fit_random_forest(formula, data, ntree = 1000, na.action = na.omit,
  importance = TRUE, seed = NULL, ...)
}
\arguments{
\item{formula}{A formula for the random forest such as y~x.}

\item{data}{Data to be used.}

\item{ntree}{Number of trees to grow. This should not be set to too small.}

\item{na.action}{A function to specify the action to be taken if NAs are found}

\item{importance}{Should importance of predictors be assessed? Boolean}

\item{seed}{Random seed for reproducibility purposes.}

\item{...}{Other arguments passed to the randomForest::randomForest function}
}
\description{
Fit a random forest model
}

