% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_coordinates}
\alias{fix_coordinates}
\title{Check existence of a variable inside a given .nc file.
If the variable does not exist, use the replacement one.}
\usage{
fix_coordinates(nc, coordinates, spare_coordinates)
}
\arguments{
\item{nc}{file with extension .nc already loaded in the R environment.}

\item{coordinates}{Grid variables (for instance, longitude and latitude).}

\item{spare_coordinates}{Spare names for coordinates. Variables such as longitude and latitude may be named differently in every
.nc file. In order to account this possibility, you can provide a set of spare names for both coordinates.}
}
\value{
Returns existing coordinates (either coordinates or spare_coordinates)
}
\description{
Check existence of a variable inside a given .nc file.
If the variable does not exist, use the replacement one.
}
