% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{format_lon_lat}
\alias{format_lon_lat}
\title{Fix latitude and longitude}
\usage{
format_lon_lat(dataframe, reference_df, variable)
}
\arguments{
\item{dataframe}{a dplyr dataframe}

\item{reference_df}{reference dataframe containing lon, lat, id_pixel, and groups.}

\item{variable}{variable to be matched for each pair of longitude and latitude coordinates}
}
\description{
A brute force approach is used to match latitude and longitude pairs and the corrisponding value of
variable.
}

