% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{get_variable_importance}
\alias{get_variable_importance}
\title{Get variable importance}
\usage{
get_variable_importance(rf_model)
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest.}
}
\description{
Get variable importance
}
