% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolating_functions.R
\name{interpolate_grid}
\alias{interpolate_grid}
\title{Load nc files to be resized. Subfunction}
\usage{
interpolate_grid(data_list, reference_df, observation_id = "id_date")
}
\arguments{
\item{data_list}{a list of dplyr dataframe to interpolate}

\item{reference_df}{a reference dataframe containing at least the following variables: longitude, latitude, id_pixel}

\item{observation_id}{id of the observation: id_date for daily observations, month for monthly observations.}
}
\value{
a dplyr dataframe
}
\description{
Load nc files to be resized. Subfunction
}

