% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolating_functions.R
\name{interpolate_single_grid}
\alias{interpolate_single_grid}
\title{Interpolate each data grid.
This function interpolates every datagrid day by day (or month by month in case
of monthly observations)}
\usage{
interpolate_single_grid(df, unique_id = "id_date", variable,
  coordinates_names, lon_range, lat_range, step)
}
\arguments{
\item{df}{a dplyr dataframe to interpolate}

\item{unique_id}{unique id of the observation (usually id_date or id_month)}

\item{variable}{variable to interpolate}

\item{coordinates_names}{names of the coordinates (longitude and latitude)}

\item{lon_range}{longitude range (min, max)}

\item{lat_range}{latitude range (min, max)}

\item{step}{step between longitude and latitude vectors}
}
\value{
a dplyr dataframe
}
\description{
Interpolate each data grid.
This function interpolates every datagrid day by day (or month by month in case
of monthly observations)
}

