% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolating-functions.R
\name{interpolate_single_grid_multi_day}
\alias{interpolate_single_grid_multi_day}
\title{Interpolate each data grid.
This function interpolates every datagrid day by day (or month by month in case
of monthly observations)}
\usage{
interpolate_single_grid_multi_day(df, unique_id, variable, coordinates_names,
  lon_range, lat_range, step, date_variable_name)
}
\arguments{
\item{df}{a dplyr dataframe to interpolate}

\item{unique_id}{unique id of the observation (usually id_date or month). Do NOT use id_pixel.}

\item{variable}{variable to interpolate}

\item{coordinates_names}{names of the coordinates (longitude and latitude)}

\item{lon_range}{longitude range (min, max)}

\item{lat_range}{latitude range (min, max)}

\item{step}{step between longitude and latitude vectors}

\item{date_variable_name}{name of the date variable. Character. Defaults to "date"}
}
\value{
a dplyr dataframe
}
\description{
Interpolate each data grid.
This function interpolates every datagrid day by day (or month by month in case
of monthly observations)
}
