% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{join_data}
\alias{join_data}
\title{Join data together}
\usage{
join_data(multiple_year_data, single_year_data, m_year_join_by = c("lon",
  "lat", "year"), s_year_join_by = c("lon", "lat"))
}
\arguments{
\item{multiple_year_data}{a list of lists of dataframes. Each dataframe should contain observations of a single variable in a certain year.}

\item{single_year_data}{a list of dplyr dataframes.}

\item{m_year_join_by}{variables to join by (multiple year dataframes)}

\item{s_year_join_by}{variables to join by (single_year_dataframes)}
}
\description{
Join multiple year data by lon, lat and year
Then join single year data by lon and lat
}
