% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_loaded_data}
\alias{join_loaded_data}
\title{This function is used for joining data together after an id pixel has been assigned
to the data coming from the interpolation process.}
\usage{
join_loaded_data(df_list, by = c("lon", "lat", "id_pixel"))
}
\arguments{
\item{df_list}{list of loaded dataframes. A list of dplyr dataframe to join together.}

\item{by}{variables to join by}
}
\value{
Returns a dplyr dataframe
}
\description{
This function is used for joining data together after an id pixel has been assigned
to the data coming from the interpolation process.
}

