% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{keep_pixels_with_group}
\alias{keep_pixels_with_group}
\title{Filter out pixels that have no group assigned}
\usage{
keep_pixels_with_group(data, group_name = "gruppo")
}
\arguments{
\item{data}{a dplyr dataframe}

\item{group_name}{name of the variable containing the group information}
}
\description{
Filter out pixels that have no group assigned
}

