% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{load_a_single_csv}
\alias{load_a_single_csv}
\title{Load a single .csv file and output a dplyr dataframe}
\usage{
load_a_single_csv(file_path, ...)
}
\arguments{
\item{file_path}{path to .csv file to load.}

\item{...}{other arguments to be used by the read.csv function}
}
\description{
Load a single .csv file and output a dplyr dataframe
}

