% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data_functions.R
\name{load_all}
\alias{load_all}
\title{Load a bunch of .nc files in a local path}
\usage{
load_all(path, variables = c("CHL1_mean"), coordinates = c("lon", "lat"),
  date_format = "ymd")
}
\arguments{
\item{path}{Path where to .nc files are located. Eg: /home/data}

\item{variables}{variables to be retrieved from the file. A character vector of length n.}

\item{coordinates}{longitude and latitude}
}
\value{
A list of all the .nc files loaded
}
\description{
Load a bunch of .nc files in a local path
}
\examples{

}

