% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{load_all_csv}
\alias{load_all_csv}
\title{Load a bunch of csv files into a single list}
\usage{
load_all_csv(main_folder_path, folder, starts_with, years = c("2011", "2012"),
  npixels = 5520, must_have_variables = c("lon", "lat"))
}
\arguments{
\item{main_folder_path}{path to the main folder}

\item{folder}{name of the folder containing the .csv files to load. A character.}

\item{starts_with}{initial part of the name of the .csv file. Each file is assumed to have a recurrent pattern
in its name where the only changing part is assumed to be the year. A character}

\item{years}{years to load. A character vector.}

\item{must_have_variables}{a vector of variables that must be included in each .csv file. Character vector.}
}
\description{
Load a bunch of csv files into a single list
}

