% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data_from_nc}
\alias{load_data_from_nc}
\title{Extract variables from a .nc file}
\usage{
load_data_from_nc(var, nc)
}
\arguments{
\item{var}{variable to be extracted}

\item{nc}{file with extension nc}
}
\value{
The requested variable
}
\description{
Extract variables from a .nc file
}

