% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-functions.R
\name{load_mask}
\alias{load_mask}
\title{Load the mask file as a data frame.}
\usage{
load_mask(mask_file = "mask_25km.nc", lon_lat_names = c("lon", "lat"),
  mask_name = "lsmask")
}
\arguments{
\item{mask_file}{path to nc file containing the binary mask}

\item{lon_lat_names}{names of coordinates in the mask file. Defaults to c("lon", "lat")}

\item{mask_name}{mask variable name}
}
\value{
a dplyr dataframe
}
\description{
Load the mask file as a data frame.
}
