% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data_functions.R
\name{load_nc_file}
\alias{load_nc_file}
\title{Load a single .nc file}
\usage{
load_nc_file(file_path, variables = c("CHL1_mean"), coordinates = c("lon",
  "lat"), date_format = "ymd")
}
\arguments{
\item{file_path}{the path of the file to load. Character.}

\item{variables}{variables to be retrieved from the file. A character vector of length n.}

\item{coordinates}{longitude and latitude}

\item{current_date}{the date of the observations. A lubridate object}
}
\value{
A dplyr dataframe
}
\description{
Load a single .nc file
}
\examples{

}

