% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_variable_from_nc}
\alias{load_variable_from_nc}
\title{Extract a single variable from a .nc file}
\usage{
load_variable_from_nc(variable, nc)
}
\arguments{
\item{variable}{variable to be extracted. Character.}

\item{nc}{file with .nc extension}
}
\value{
The requested variable. Data structure depends on which variable is recovered.
}
\description{
This function extracts the variable var from a given nc file.
}
