% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_path}
\alias{make_path}
\title{Make path function}
\usage{
make_path(filename, path)
}
\arguments{
\item{filename}{name of the file}

\item{path}{path to folder}
}
\value{
full path to a file
}
\description{
Work in progress.
}

