% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{multiplot}
\alias{multiplot}
\title{Fit multiple plots in a single plot}
\usage{
multiplot(plots = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plots}{a list of ggplot objects}

\item{cols}{number of columns to display the plots}

\item{layout}{layout}
}
\description{
Fit multiple plots in a single plot
}
