% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_x-data.R
\docType{data}
\name{new_x}
\alias{new_x}
\title{Example of reference dataframe for interpolation and join.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5520 rows and 4 columns.}
\usage{
data(new_x)
}
\description{
Contains longitude (lon), latitude (lat), id pixel (id_pixel) and group (gruppo)
}
\examples{
data(new_x)

}
\keyword{datasets}
