% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{optimal_clusters_number}
\alias{optimal_clusters_number}
\title{###############################################################################
Select the optimal number of clusters using the Calinski-Harabasz index}
\usage{
optimal_clusters_number(x, minC = 2, maxC = 10, plot_show = FALSE)
}
\arguments{
\item{x}{a dataframe or a matrix containing the data. The data should be standardized for better results}

\item{minC}{minumum number of clusters Numeric. (Integer)}

\item{maxC}{maximum number of clusters. Numeric. (Integer)}

\item{plot_show}{plot the results. Boolean.}
}
\value{
The optimal number of clusters. (The number of clusters for which the index calculated is maximum)
}
\description{
This function is used to standardize the data before running the kmeans analysis function.
The output of this function is a dataframe ready to be used for the kmeans analysis.
}

