% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{partial_dependence_plot}
\alias{partial_dependence_plot}
\title{Partial dependence plot}
\usage{
partial_dependence_plot(rf_model, data)
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest.}

\item{data}{data used to fit the random forest model}
}
\description{
Partial dependence plot
}

