% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{plot_density_imputed_na}
\alias{plot_density_imputed_na}
\title{Plot the density of imputed NAs vs actual data}
\usage{
plot_density_imputed_na(data)
}
\arguments{
\item{data}{dataset imputed obtained from the approximate_NAs function,
must have an "imputed_dataset" attribute.}
}
\description{
Density plot of imputed NAs vs actual data.
}
