% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{plot_error_vs_trees}
\alias{plot_error_vs_trees}
\title{Plot model error vs number of trees}
\usage{
plot_error_vs_trees(rf_model, title)
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest.}

\item{title}{plot title}
}
\description{
Plot model error vs number of trees
}

