% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{plot_results}
\alias{plot_results}
\title{Plot results.}
\usage{
plot_results(kmeans_results)
}
\arguments{
\item{kmeans_results}{a list of results obtained from the function kmeans_analysis}
}
\description{
This function is used to plot the number of clusters against the Calinski-Harabasz index
}
