% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{ppdplot}
\alias{ppdplot}
\title{Plot partial dependence plots. A revised version of the official function.
The function had to be revised due to the use of NSE on the x.var argument.}
\usage{
ppdplot(x, pred.data, x.var, which.class, w, plot = TRUE, add = FALSE,
  n.pt = min(length(unique(pred.data[, xname])), 51), rug = TRUE,
  xlab = deparse(substitute(x.var)), ylab = "",
  main = paste("Partial Dependence on", deparse(substitute(x.var))), ...)
}
\arguments{
\item{x}{altro}

\item{pred.data}{altro}

\item{x.var}{altro}

\item{which.class}{altro}

\item{w}{altro}

\item{plot}{altro}

\item{add}{altro}

\item{n.pt}{altro}

\item{rug}{altro}

\item{xlab}{altro}

\item{ylab}{altro}

\item{main}{altro}

\item{...}{altro}
}
\description{
Plot partial dependence plots. A revised version of the official function.
The function had to be revised due to the use of NSE on the x.var argument.
}
