% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{raw_prediction_plot}
\alias{raw_prediction_plot}
\title{Plot OOB prediction estimates}
\usage{
raw_prediction_plot(rf_model)
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest}
}
\description{
Extract the OOB prediction estimates from the random forest and plot them.
}

