% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics-calculation-functions.R
\name{reshape_statistics}
\alias{reshape_statistics}
\title{Given a list of variables and a dataframe, make a reshaped (from wide to long)
dataframe for each variable and output a list of dataframes}
\usage{
reshape_statistics(data, id = "id_pixel", key = "id_date",
  other_id = list("lon", "lat"))
}
\arguments{
\item{data}{dplyr dataframe to be reshaped containing the variables to be used in the reshaping process}

\item{id}{unique identifier for each pixel}

\item{key}{key argument in tidyr::spread}

\item{other_id}{other unique identifiers (for instance: longitude and latitude)}
}
\description{
Given a list of variables and a dataframe, make a reshaped (from wide to long)
dataframe for each variable and output a list of dataframes
}

