% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-functions.R
\name{resize_mask}
\alias{resize_mask}
\title{Resize the mask}
\usage{
resize_mask(mask, reference_df, step = 0.25, coordinates_names = c("lon",
  "lat"), plot_ = FALSE)
}
\arguments{
\item{mask}{a mask dataframe loaded using the load_mask function.}

\item{reference_df}{a reference dataframe containing at least the following variables: longitude, latitude, id_pixel.}

\item{step}{step (resolution) in the longitude-latitude grid. Set to 0.25 by default.}

\item{coordinates_names}{names of the coordinates (longitude and latitude).}

\item{plot_}{Should the function plot the original and interpolated map for a visual comparison? Boolean. FALSE by default.}
}
\value{
a dplyr dataframe
}
\description{
Resize the mask
}
