% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_filenames_by_date}
\alias{select_filenames_by_date}
\title{This function basically handles how to proceed when selecting filenames by date}
\usage{
select_filenames_by_date(file_names, from, to, date_format)
}
\arguments{
\item{file_names}{a list of file names}

\item{from}{Initial date. A character or NULL}

\item{to}{Ending date. A character or NULL}

\item{date_format}{date format}
}
\value{
Returns a list of selected filenames to be loaded
}
\description{
Select names of the file to be loaded according to boundary conditions.
4 possible cases:
1. to = from = NULL. No dates have been provided, all the file_names in file_names are returned
2. to = NULL, from = given. All files from the given date onwards are selected
3. to = given, from = NULL. All files up to the given date are selected
4. to = from = given. Files within the dates interval are selected
}
