% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{shift_lon_lat}
\alias{shift_lon_lat}
\title{Shift lon and lat}
\usage{
shift_lon_lat(dataframe, reference_dataframe)
}
\arguments{
\item{dataframe}{a dplyr dataframe}

\item{reference_dataframe}{reference dataframe containing lon, lat, id_pixel, and groups.}
}
\description{
Simply replace lon and lat with those in the reference dataframe
}

