% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{standardize_data}
\alias{standardize_data}
\title{###############################################################################
Standardize data before running kmeans}
\usage{
standardize_data(data, exclude_variables = list("lon", "lat", "id_pixel"))
}
\arguments{
\item{data}{a dplyr dataframe containing the data to be standardized}

\item{exclude_variables}{variables to be excluded from the standardization process. A list of characters.}
}
\value{
A matrix
}
\description{
This function is used to standardize the data before running the kmeans analysis function.
This function removes NAs too.
The output of this function is a dataframe ready to be used for the kmeans analysis.
}

