% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{variable_importance_plot}
\alias{variable_importance_plot}
\title{Plot variable importance}
\usage{
variable_importance_plot(rf_model)
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest.}
}
\description{
Plot variable importance
}

