% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_calculation_functions.R
\name{aggregate_statistics}
\alias{aggregate_statistics}
\title{Calculate a series of stastistics on aggregated data}
\usage{
aggregate_statistics(data, variable = "CHL1_mean", stat_funs = list(avg =
  "mean(x, na.rm=TRUE)"), groups = list("id_pixel", "id_date"),
  unique_id = list("lat", "lon", "id_pixel"))
}
\arguments{
\item{data}{a dplyr data frame}

\item{variable}{variable to use in the calculation of the statistics}

\item{stat_funs}{a list of functions to calculate the statistics. Each function must be expressed either as a
formula or as a character. For instance, if I would like to calculate the mean, I could set stat_funs = list(avg = "mean(x)").
You can calculate multiple statistics too, for instance, to calculate both the mean and the standard deviation, just set
stat_funs = list(avg = "mean(x)", sd = "sd(x)"). Note that:
1. The name of the function inside the list will be the name of the column of the statistic calculated.
2. The variable inside each function should be expressed as "x".
3. Each function must accept a vector as an input and output a single number.}

\item{groups}{variables to group by (ie to aggregate by). A list.}

\item{unique_id}{A list of unique identifiers.}
}
\value{
A dplyr data frame.
}
\description{
This function groups the observations and calculates a series of statistics.
}

