% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_date_from_filepath}
\alias{extract_date_from_filepath}
\title{Extract date from filepath}
\usage{
extract_date_from_filepath(path, date_format = "ymd")
}
\arguments{
\item{path}{path to file}

\item{date_format}{Date format. Should match a lubridate}
}
\value{
full path to a file
}
\description{
Assumptions:
1. Date is in the following format: dddddddd where d is an integer.
2. Date is the first 8 digits matched in the filename.
}
\details{
Notes:
Date format can be set
}

