% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{format_lon_lat}
\alias{format_lon_lat}
\title{Fix latitude and longitude}
\usage{
format_lon_lat(dataframe, reference_df, variable, lon_lat_names = c("lon",
  "lat"))
}
\arguments{
\item{dataframe}{a dplyr dataframe.}

\item{reference_df}{a reference dataframe that must contain at least longitude and latitude. A dataframe.}

\item{variable}{variable to be used in the reformatting process. The values of this variable will
be reordered according to the new longitude and latitude values. Character.}

\item{lon_lat_names}{name of longitude and latitude variables. A character vector.
Example: by default lon_lat_names = c("lon", "lat")}
}
\description{
A brute force approach is used to match latitude and longitude pairs and the corrisponding value of
variable.
}

