% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-random-forest-data.R
\name{format_lon_lat_list}
\alias{format_lon_lat_list}
\title{Fix latitude and longitude}
\usage{
format_lon_lat_list(df_list, reference_df, variable = NULL,
  reformat = FALSE, shift = FALSE)
}
\arguments{
\item{df_list}{a list of dplyr dataframe}

\item{reference_df}{reference dataframe containing lon, lat, id_pixel, and groups.}

\item{variable}{variable to be used in the reformatting process}

\item{reformat}{Should lon and lat be fixed from scratch? Boolean.}

\item{shift}{Should lon and lat be shifted? Boolean}
}
\description{
Align longitude and latitude values.
}
\details{
If reformat = TRUE a "brute force" reformat is performed to match those in the reference_df.
If shift = TRUE the values of longitude and latitude are assumed to be shifted by
a negligible amount and replaced with those in the reference dataframe.
}

