% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-analysis.R
\name{kmeans_analysis}
\alias{kmeans_analysis}
\title{###############################################################################
Run kmeans analysis}
\usage{
kmeans_analysis(x, n_centers, random_seed, ...)
}
\arguments{
\item{x}{a dataframe or a matrix containing the data. The data should be standardized for better results}

\item{n_centers}{Number of clusters to be used}

\item{random_seed}{Random seed set for reproducibility purposes. Numeric. (Integer)}

\item{...}{Other arguments accepted by the stats::kmeans function.}
}
\value{
An object of class kmeans
}
\description{
Note: this function takes as argument all the arguments used by the stats::kmeans function. Simply
pass the extra arguments after the x, n_ceters and random_seed arguments.
}

