% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data_functions.R
\name{load_all_as_list}
\alias{load_all_as_list}
\title{Load all .nc files in a given local path}
\usage{
load_all_as_list(path, variables = c("CHL1_mean"), coordinates = c("lon",
  "lat"), spare_coordinates = c("longitude", "latitude"),
  date_format = "ymd", date_match_position = 1)
}
\arguments{
\item{path}{Path where to .nc files are located. Eg: /home/data. Character.}

\item{variables}{variables to be retrieved from the file. A character vector of length n.}

\item{coordinates}{Grid variables (for instance, longitude and latitude).}

\item{spare_coordinates}{Spare names for coordinates. Variables such as longitude and latitude may be named differently in every
.nc file. In order to account this possibility, you can provide a set of spare names for both coordinates.}

\item{date_format}{date format. By default it is set to be "ymd" (year, month, day). This is the format of the date in each
filename. It can be set to other values such as "dmy" and "mdy". Check lubridate's package help for more information.}

\item{date_match_position}{An integer. If the name of each file contains one or more dates, chose which one will be used as the
current date for the file. Example: if a file is named "20150202file2_20150706.nc", if date_match_position is set
to 1 (default), the first date, 20150202 will be used. If you'd like to use the second one, set the parameter to 2 and so on.}
}
\value{
A list of all the .nc files loaded
}
\description{
Load all .nc files in a given local path
}
\examples{
# Load all .nc files in /home/data, extract the variable "CHL1_mean" and use longitude and latitude to uniquely identify each observation
# loaded_files_list <- load_all_as_list(path = /home/data, variables = c("CHL1_mean"), coordinates = c("lon","lat"))

}

