% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data-to-be-resized.R
\name{load_nc_to_resize}
\alias{load_nc_to_resize}
\title{Load all .nc files to be resized as a list in a given local path}
\usage{
load_nc_to_resize(path, from = NULL, to = NULL, variables = c("qnet"),
  coordinates = c("lon", "lat"), spare_coordinates = c("longitude",
  "latitude"), time_variable = "time", lower_left_lat_lon = c(52, -65),
  upper_right_lat_lon = c(67, -42), monthly = FALSE)
}
\arguments{
\item{path}{Path where to .nc files are located. Example: /home/data. Character.}

\item{from}{starting year (included). Either a numeric or a character. Example: 2009}

\item{to}{ending year(included). Either a numeric or a character. Example: 2010}

\item{variables}{variables to be extracted from .nc file}

\item{coordinates}{longitude and latitude names}

\item{spare_coordinates}{Spare names for coordinates. Variables such as longitude and latitude may be named differently in every
.nc file. In order to account this possibility, you can provide a set of spare names for both coordinates. Set by default
to be: c("longitude","latitude").}

\item{time_variable}{variable representing the frequency of the observations. Character.}

\item{lower_left_lat_lon}{lower left corner latitude and longitude of the selected area.}

\item{upper_right_lat_lon}{upper right corner latitude and longitude of the selected area.}

\item{monthly}{whether data has a monthly or an annual frequency. TRUE if data has a monthly frequency.
FALSE if frequency is annual. FALSE by default.}
}
\value{
a list of dplyr dataframes
}
\description{
Note: these files contain a time variable. They represent observations of a
certain variable with a given frequency. The variable "time" defines the name of
the sampling frequency variable. The function assumes that the data has a daily
frequency. If the frequency is monthly, set the monthy variable to TRUE.
}

