% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_path}
\alias{make_path}
\title{Make path function}
\usage{
make_path(file_name, path)
}
\arguments{
\item{file_name}{name of the file. Character}

\item{path}{path to folder. Character}
}
\value{
full path to a file. Character
}
\description{
Given a path to a folder and the name of a file in that folder, this function
builds the path to that file.
}
\details{
Example:
path = "/home/user/date"
filename = "data.txt"
Return = "/home/user/date/data.txt"
}

