% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{partial_dependence_plot}
\alias{partial_dependence_plot}
\title{Partial dependence plot}
\usage{
partial_dependence_plot(rf_model, data, show_plots = FALSE, cols = 2,
  ylabel = "y", verbose = FALSE)
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest.}

\item{data}{data used to fit the random forest model}

\item{show_plots}{should plots be showed? Boolean}

\item{cols}{number of columns to use when plotting the dependence plots. Defaults to 2.}

\item{ylabel}{label of the y axis. Character. Defaults to "y".}

\item{verbose}{be verbose? Boolean.}
}
\value{
a list with x and y values for the partial dependence graph for each variable. If show_plots
is TRUE, NULL is returned.
}
\description{
Partial dependence plot
}

