% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{predictive_map}
\alias{predictive_map}
\title{Plot predictive map}
\usage{
predictive_map(rf_model, data)
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest.}

\item{data}{data used to fit the random forest model}
}
\description{
Plot predictive map
}

