% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest-model.R
\name{predictive_map}
\alias{predictive_map}
\title{Plot predictive map}
\usage{
predictive_map(rf_model, data, facet_plot = FALSE, facet_by = "year",
  lon_lat_names = list("lon", "lat"))
}
\arguments{
\item{rf_model}{random forest model obtained from the function fit_random_forest.}

\item{data}{data used to fit the random forest model}

\item{facet_plot}{Should the plot be categorized? If TRUE, then the argument "facet_by" must be set. If FALSE,
an average prediction is computed and plotted. Boolean}

\item{facet_by}{variable to use in facet plots. Character.}

\item{lon_lat_names}{names of latitude and longitude in the dataset. A list of characters. Defaults to list("lon", "lat")}
}
\description{
Plot predictive map
}
