% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data-to-be-resized.R
\name{recover_nc_data}
\alias{recover_nc_data}
\title{Load nc files to be resized. Subfunction}
\usage{
recover_nc_data(file_path, variables, coordinates, spare_coordinates,
  time_variable)
}
\arguments{
\item{file_path}{Path where the file is located. Example: /home/data. Character.}

\item{variables}{variables to be extracted from .nc file}

\item{coordinates}{longitude and latitude names}

\item{spare_coordinates}{Spare names for coordinates. Variables such as longitude and latitude may be named differently in every
.nc file. In order to account this possibility, you can provide a set of spare names for both coordinates. Set by default
to be: c("longitude","latitude").}

\item{time_variable}{variable representing the frequency of the observations. Character.}
}
\value{
a dplyr dataframe
}
\description{
Load nc files to be resized. Subfunction
}

