% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-data-functions.R
\name{reshape_data}
\alias{reshape_data}
\title{Reshape the raw data and output into a dplyr dataframe.}
\usage{
reshape_data(raw_data, variables, expand_variables, current_date)
}
\arguments{
\item{raw_data}{raw data extracted using load_nc_file. A list.}

\item{variables}{variables to be retrieved from the file. A character vector of length n.}

\item{expand_variables}{variables to be used as x and y reference of the image grid. These variables must be included in the raw_data list.}

\item{current_date}{The date of the observation.}
}
\value{
A dplyr dataframe
}
\description{
Reshape the raw data and output into a dplyr dataframe.
}

